Improved Transport sync performance
