+++ services/syra-daw-fe/src/ui/molecules/BackboneMixer/BackboneAudioMixer.tsx
@@ -5,6 +5,7 @@ import { channelStore } from '../../../recoil/channelStore';
 import { useRecoilValue } from 'recoil';
 import useConnectMidiWithInstrumentChannel from '../../../hooks/audio/useConnectMidiWithInstrumentChannel';
 import { ChannelType } from '../../../types/Channel';
+import useSyncTransportToSoul from "../../../hooks/soul/useSyncTransportToSoul";
 
 interface Props {
   channelId: string;
@@ -23,6 +24,8 @@ function BackboneAudioMixer({channelId}: Props) {
   const soulInstance = useRecoilValue(channelStore.soulInstance(channelId));
   const plugins = useRecoilValue(channelStore.findActivePluginsByIds(pluginIds));
   const [connect, disconnect] = useConnectMidiWithInstrumentChannel(channelId);
+
+  useSyncTransportToSoul();
 
   useEffect(() => {
     // TODO: WE COULD PROBABLY JUST CREATE A INSTANCE LIKE Tone.Solo, BUT FOR MUTING. THIS WOULD SAVE US A REWIRING CALL.
@@ -44,8 +47,6 @@ function BackboneAudioMixer({channelId}: Props) {
   }, [isSolo]);
 
   useEffect(() => {
-    console.log('run effect');
-
     mixerChannelStrip.disconnect(soulInstance?.audioNode);
     mixerChannelStrip.rewireAudio(plugins.map(plugin => plugin.audioNode), soulInstance?.audioNode)
   }, [soulInstance, plugins]);
